Pass version along to docker images

+++ src/main/php/xp/lambda/CreateRuntime.class.php
@@ -6,24 +6,39 @@ use util\cmd\Console;
 class CreateRuntime {
   use Docker;
 
-  private $target, $rebuild;
+  private $version, $target, $rebuild;
 
-  public function __construct(Path $target, bool $rebuild= false) {
-    $this->target= $target;
+  public function __construct(string $version, Path $target, bool $rebuild= false) {
+    $this->version= $version;
+    $this->target= new Path(sprintf($target, $version));
     $this->rebuild= $rebuild;
   }
 
   public function run(): int {
     $docker= $this->command();
-    $runtime= $this->image('runtime', [], $this->rebuild);
-    $container= uniqid();
 
+    // Build test and runtime containers when -b is given
+    if ($this->rebuild) {
+      $runtime= $this->image('test', $this->version, ['runtime' => []], true)['runtime'];
+    } else {
+      $runtime= $this->image('runtime', $this->version)['runtime'];
+    }
+    if (null === $runtime) return 1;
+
+    // Verify runtime
+    $code= "echo ' => PHP ', PHP_VERSION, ' & Zend ', zend_version(), ' @ ', php_uname(), PHP_EOL, ' => ', implode(' ', get_loaded_extensions()), PHP_EOL;";
+    Console::writeLine();
+    Console::writeLine("[+] Running {$runtime}\e[34m");
+    passthru("{$docker} run --rm {$runtime} /opt/php/bin/php -r \"{$code}\"", $result);
+    Console::writeLine("\e[0m");
+
+    // Extract runtime
+    $container= uniqid();
     $commands= [
       "{$docker} create --name {$container} {$runtime}",
       "{$docker} cp {$container}:/opt/php/runtime.zip {$this->target}",
       "{$docker} rm -f {$container}",
     ];
-
     Console::writeLine('[+] Creating ', $this->target);
     foreach ($commands as $command) {
       Console::writeLinef("\e[34m => %s\e[0m", $command);
